+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
-## [v0.0.1] - 2024-11-14
+## [v0.0.2] - 2024-11-19
+
+### Added
+
+- Color picker for GoodSpecifiction box color
+- Localization translator command (`beta`)
+- Hide unity .meta files command
+- Project setup command
+- Extension pack ([`Unity`](https://marketplace.visualstudio.com/items?itemName=VisualStudioToolsForUnity.vstuc), [`.NET Install Tool`](https://marketplace.visualstudio.com/items?itemName=ms-dotnettools.vscode-dotnet-runtime), [`C#`](https://marketplace.visualstudio.com/items?itemName=ms-dotnettools.csharp) & [`C# Dev Kit`](https://marketplace.visualstudio.com/items?itemName=ms-dotnettools.csdevkit))
+
+### Changed
+
+- Better csv snippets
+- Moved commands to seperate files
+
+## [v0.0.1] - 2024-11-18
 
 ### Added